agents/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.91.0
Flask==3.1.0
mcp[cli]==1.7.1
google-adk==0.4.0
python-dateutil==2.9.0.post0
humanize==4.12.3
nest_asyncio==1.6.0
asyncclick==8.1.8.0
a2a_common-0.1.0-py3-none-any.whl
deprecated==1.2.18
psycopg2-binary 
python-dateutil
python-dotenv
requests

